test: consolidate 8 per-module io tests into 2; cut boilerplate; fix fastfarm import
Aggressive reduction (the strengthened driver round-trip + differential harness now cover
read/write/OutList integration, making per-module boilerplate redundant):
- 8 test_io_*.py -> test_io_onshore.py + test_io_offshore.py, keeping only edge cases
  (HydroDyn NBodyMod matrices, AeroDyn polars/coords, SubDyn GuyanDamp, StC, ExtPtfm
  superelement, BeamDyn geometry, HAWC wind, etc.).
- Cut all is_module_io/returns_dict/per-module-roundtrip/trivial-scalar reads; the 13
  subclass checks collapse to one parametrized test in test_io_base.py.
- Fixed a pre-existing broken import in test_driver_fastfarm.py (FASTFarmDriver was
  imported from the package, not openfast_io.drivers.fastfarm) — the whole fast suite
  couldn't even collect; unnoticed because CI only runs test_of_io_pytest.py.

Fast suite: 141 passed, 0 failed. CI (ctest -L openfast_io) still runs only
test_of_io_pytest.py — unchanged.
